Moved notification permission to when the user joins event.
